# E2E Performance test to print latency numbers for vsphere volume lifecycle operations