--- a/test/e2e/storage/BUILD
+++ b/test/e2e/storage/BUILD
@@ -34,6 +34,7 @@ go_library(
         "vsphere_volume_fstype.go",
         "vsphere_volume_master_restart.go",
         "vsphere_volume_ops_storm.go",
+        "vsphere_volume_perf.go",
         "vsphere_volume_placement.go",
         "vsphere_volume_vsan_policy.go",
     ],