@pshahzeb 2017-11-07
@jeffvance 
Thank you for your patient review.

I have addressed all the comments in latest commit.